+++ bosesoundtouchapi/models/preset.py
@@ -254,6 +254,37 @@ class Preset(SoundTouchModelRequest):
             self._UpdatedOn = value
 
 
+    def ToDictionary(self, encoding:str='utf-8') -> dict:
+        """
+        Returns a dictionary representation of the class.
+        
+        Args:
+            encoding (str):
+                encode type (e.g. 'utf-8', 'unicode', etc).  
+                Default is 'utf-8'.
+        """
+        if encoding is None:
+            encoding = 'utf-8'
+            
+        contentItem:dict = {}
+        if self._ContentItem is not None:
+            contentItem = self._ContentItem.ToDictionary(encoding)
+        
+        result:dict = {}
+        
+        if self._PresetId is not None: 
+            result['PresetId'] = self._PresetId
+        if self._CreatedOn is not None: 
+            result['CreatedOn'] = self._CreatedOn
+        if self._UpdatedOn is not None: 
+            result['UpdatedOn'] = self._UpdatedOn
+        if self._SourceTitle is not None: 
+            result['SourceTitle'] = self._SourceTitle
+        result['ContentItem'] = contentItem
+
+        return result
+        
+
     def ToElement(self, isRequestBody:bool=False) -> Element:
         """ 
         Overridden.