+++ bosesoundtouchapi/models/sourcelist.py
@@ -1,7 +1,6 @@
 # external package imports.
 from typing import Iterator
 from xml.etree.ElementTree import Element, tostring
-import xmltodict
 
 # our package imports.
 from ..bstutils import export
@@ -151,8 +150,8 @@ class SourceList:
         
 
     def ToDictionary(self, encoding:str='utf-8') -> dict:
-        """ 
-        Returns a dictionary representation of the class. 
+        """
+        Returns a dictionary representation of the class.
         
         Args:
             encoding (str):
@@ -161,15 +160,15 @@ class SourceList:
         """
         if encoding is None:
             encoding = 'utf-8'
-        elm = self.ToElement()
-        xml = tostring(elm, encoding=encoding).decode(encoding)
+            
+        result:dict = {}
         
-        # convert xml to dictionary.
-        oDict:dict = xmltodict.parse(xml,
-                                     encoding=encoding,
-                                     process_namespaces=False)
-        return oDict
+        if self._DeviceId is not None: 
+            result['DeviceId'] = self._DeviceId
+        result['SourceItems'] = [ item.ToDictionary(encoding) for item in self._SourceItems ]
 
+        return result
+        
 
     def ToElement(self, isRequestBody:bool=False) -> Element:
         """