+++ CHANGELOG.md
@@ -6,6 +6,25 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.59 ] - 2024/04/25
+
+  * Removed `xmltodict` requirement.
+  * Added method `ContentItem.ToDictionary` to return a dictionary representation of the class.
+  * Added method `MediaItemContainer.ToDictionary` to return a dictionary representation of the class.
+  * Added method `NavigateItem.ToDictionary` to return a dictionary representation of the class.
+  * Updated `NavigateResponse.ToDictionary` method to remove the xmltodict requirement.
+  * Added method `Preset.ToDictionary` to return a dictionary representation of the class.
+  * Updated `PresetList.ToDictionary` method to remove the xmltodict requirement.
+  * Added method `Recent.ToDictionary` to return a dictionary representation of the class.
+  * Updated `RecentList.ToDictionary` method to remove the xmltodict requirement.
+  * Updated `SearchResponse.ToDictionary` method to remove the xmltodict requirement.
+  * Added method `SearchResult.ToDictionary` to return a dictionary representation of the class.
+  * Updated `SearchStationArtists.ToDictionary` method to remove the xmltodict requirement.
+  * Added method `SearchStationResults.ToDictionary` to return a dictionary representation of the class.
+  * Updated `SearchStationSongs.ToDictionary` method to remove the xmltodict requirement.
+  * Added method `SourceItem.ToDictionary` to return a dictionary representation of the class.
+  * Updated `SourceList.ToDictionary` method to remove the xmltodict requirement.
+
 ###### [ 1.0.58 ] - 2024/04/15
 
   * Updated `NowPlayingStatus.ContainerArtUrl` property to return the correct image url of the playing content: the `ArtUrl` value is returned if present; otheriwse the `ContentItem.ContainerArt` url is returned if present; otherwise, None is returned.  Prior to this fix, the `ContentItem.ContainerArt` value was considered first which did not always match the `ArtUrl` value and thus caused an incorrect image to be displayed for the currently playing artist and track value.